third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/reveal-execution-line-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that execution line is revealed and highlighted when debugger is paused.


Running: testRevealAndHighlightExecutionLine
Set timer for test function.
Execution line revealed and highlighted.
